feat: Explosions feedback and positive audio feedback added
This commits is about the addition of the positive and negative feedback: Including when the user obtain the right result, and when the user exceed the limit of the interactions. Other addition is to correct the interaction when the user completed the phase, at this moment the user can not interact with the wheel.